refactor: move agentcompose tests to domain packages
