+++ internal/agentcompose/capability/capability_proxy_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package capability
 
 import (
 	driverpkg "agent-compose/internal/driver"
