+++ internal/agentcompose/capability/capability_service_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package capability
 
 import (
 	"context"
