+++ internal/agentcompose/config/config_store_migration_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package config
 
 import (
 	appconfig "agent-compose/internal/config"
