+++ internal/agentcompose/events/event_dispatcher_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package events
 
 import (
 	"context"
