+++ internal/agentcompose/exec/exec_service_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package exec
 
 import (
 	"context"
