+++ internal/agentcompose/image/image_auto_backend_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package image
 
 import (
 	"context"
