+++ internal/agentcompose/image/image_service_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package image
 
 import (
 	"context"
