+++ internal/agentcompose/llm/llm_facade_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package llm
 
 import (
 	"bytes"
