+++ internal/agentcompose/loader/loader_schedule_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package loader
 
 import (
 	"strings"
