+++ internal/agentcompose/loader/loader_timer_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package loader
 
 import (
 	appconfig "agent-compose/internal/config"
