+++ internal/agentcompose/session/session_rpc_bridge_test.go
@@ -1,4 +1,4 @@
-package agentcompose
+package session
 
 import (
 	appconfig "agent-compose/internal/config"
